Added components "NewGame" and "UpdateGame" Deleted component "GameDetails"
